Заменил числовое представление с double на decimal issue #42

+++ src/ScriptEngine/Machine/Contexts/COMWrapperContext.cs
@@ -83,7 +83,7 @@ namespace ScriptEngine.Machine.Contexts
             }
             else if (type == typeof(double))
             {
-                return ValueFactory.Create((double)objParam);
+                return ValueFactory.Create((decimal)objParam);
             }
             else if (type == typeof(DateTime))
             {
